size?

+++ client/src/fileutils.c
@@ -204,7 +204,7 @@ char *newfilenamemcopyEx(const char *preferredName, const char *suffix, savePath
     char *pfn = fileName;
 
     // user preference save paths
-    int save_path_len = path_size(e_save_path);
+    size_t save_path_len = path_size(e_save_path);
     if (save_path_len) {
         snprintf(pfn, len, "%s%s", g_session.defaultPaths[e_save_path], PATHSEP);
         pfn += save_path_len + strlen(PATHSEP);
